Reach minimum 80% line test coverage for all files

+++ src/controllers/venue.controller.ts
@@ -4,19 +4,19 @@ import mongoose from 'mongoose';
 import { Venue as VenueInterface } from '../interfaces/Venue.interface';
 import Venue from '../models/Venue.model';
 import { mapEmoji } from '../utils/venue.utils';
+import { ObjectId } from 'mongodb';
 
 export const createVenue = async (req: Request, res: Response) => {
   const newVenue = new Venue(req.body);
 
   try {
     await newVenue.save();
-  } catch (error: any) {
-    console.log(error);
-    return res.status(500).send({ message: error.message });
-  } finally {
+
     return res
       .status(201)
       .send({ message: 'Successfully created venue!', venue: newVenue });
+  } catch (error: any) {
+    return res.status(500).send({ message: error.message });
   }
 };
 
@@ -24,6 +24,9 @@ export const getVenue = async (req: Request, res: Response) => {
   let targetVenue: VenueInterface[] = [];
 
   try {
+    if (!ObjectId.isValid(req.params?.venueId)) {
+      return res.status(400).send({ message: 'Invalid venue ID!' });
+    }
     const venueId = req.params?.venueId;
     const userId = req.query?.userId;
     targetVenue = await Venue.aggregate([
@@ -79,27 +82,37 @@ export const getVenue = async (req: Request, res: Response) => {
         },
       },
     ]);
-  } catch (error: any) {
-    console.log(error);
-    return res.status(500).send({ message: error.message });
-  } finally {
+
+    if (targetVenue.length == 0) {
+      return res.status(400).send({ message: 'Venue does not exist!' });
+    }
+
     return res
       .status(200)
       .send({ message: 'Successfully found venue!', venue: targetVenue[0] });
+  } catch (error: any) {
+    return res.status(500).send({ message: error.message });
   }
 };
 
 export const addReactionToVenue = async (req: Request, res: Response) => {
   try {
-    await Venue.findByIdAndUpdate(req.params?.venueId, {
+    if (!ObjectId.isValid(req.params?.venueId)) {
+      return res.status(400).send({ message: 'Invalid venue ID!' });
+    }
+
+    const result = await Venue.findByIdAndUpdate(req.params?.venueId, {
       $push: { reactions: req.body },
     });
-  } catch (error: any) {
-    return res.status(500).send({ message: error.message });
-  } finally {
+
+    if (result == undefined) {
+      return res.status(400).send({ message: 'Venue does not exist!' });
+    }
     return res
       .status(200)
       .send({ message: 'Successfully added reaction to Venue!' });
+  } catch (error: any) {
+    return res.status(500).send({ message: error.message });
   }
 };
 
@@ -108,25 +121,41 @@ export const deleteReactionFromVenue = async (req: Request, res: Response) => {
     const venueId = req.params?.venueId;
     const userId = req.query?.userId;
     const emoji = mapEmoji(req.query?.emoji as string);
-    await Venue.updateOne(
+
+    if (!ObjectId.isValid(req.params?.venueId)) {
+      return res.status(400).send({ message: 'Invalid venue ID!' });
+    }
+
+    const result = await Venue.updateOne(
       { _id: venueId },
       { $pull: { reactions: { userId: userId, emoji: emoji } } }
     );
-  } catch (error: any) {
-    return res.status(500).send({ message: error.message });
-  } finally {
+
+    if (result.modifiedCount === 0) {
+      return res.status(400).send({ message: 'Reaction not found!' });
+    }
+
     return res
       .status(200)
       .send({ message: 'Successfully deleted reaction from Venue!' });
+  } catch (error: any) {
+    return res.status(500).send({ message: error.message });
   }
 };
 
 export const deleteVenue = async (req: Request, res: Response) => {
   try {
-    await Venue.deleteOne({ _id: req.params?.venueId });
+    if (!ObjectId.isValid(req.params?.venueId)) {
+      return res.status(400).send({ message: 'Invalid venue ID!' });
+    }
+    const result = await Venue.deleteOne({ _id: req.params?.venueId });
+
+    if (result.deletedCount === 0) {
+      return res.status(400).send({ message: 'Reaction not found!' });
+    }
+
+    return res.status(200).send({ message: 'Successfully deleted venue!' });
   } catch (error: any) {
     return res.status(500).send({ message: error.message });
-  } finally {
-    return res.status(200).send({ message: 'Successfully deleted venue!' });
   }
 };
